+++ R/transform.R
@@ -139,16 +139,23 @@ st_transform.sfg = function(x, crs = st_crs(x), ...) {
 
 #' @name st_transform
 #' @param type character; one of \code{have_datum_files}, \code{proj}, \code{ellps}, \code{datum}, \code{units} or \code{prime_meridians}; see Details.
+#' @param path character; PROJ search path to be set
 #' @export
-#' @details \code{st_proj_info} lists the available projections, ellipses, datums or units supported by the Proj.4 library when \code{type} is equal to proj, ellps, datum or units; when \code{type} equals \code{have_datum_files} a boolean is returned indicating whether datum files are installed and accessible (checking for \code{conus}).
+#' @details \code{sf_proj_info} lists the available projections, ellipses, datums, units, or data search path of the PROJ library when \code{type} is equal to proj, ellps, datum, units or path; when \code{type} equals \code{have_datum_files} a boolean is returned indicating whether datum files are installed and accessible (checking for \code{conus}).
 #'
 #' PROJ >= 6 does not provide option \code{type = "datums"}. PROJ < 6 does not provide the option \code{type = "prime_meridians"}.
 #' @examples
-#' st_proj_info("datum")
-st_proj_info = function(type = "proj") {
+#' sf_proj_info("datum")
+sf_proj_info = function(type = "proj", path) {
 
 	if (type == "have_datum_files")
 		return(CPL_have_datum_files(0))
+
+	if (type == "path")
+		return(CPL_get_data_dir(FALSE))
+	
+	if (!missing(path) && is.character(path))
+		return(invisible(CPL_set_data_dir(path)))
 
     opts <- c("proj", "ellps", "datum", "units", "prime_meridians")
     if (!(type %in% opts))